+++ Source/Classes/Internal/ConcreteOverlayContainerContextTransitioning.swift
@@ -11,11 +11,13 @@ struct ConcreteOverlayContainerContextTransitioning:
     OverlayContainerContextTransitioning,
     OverlayContainerTransitionCoordinatorContext {
 
+    let isCancelled: Bool
+    let isAnimated: Bool
     let overlayViewController: UIViewController
     let overlayTranslationHeight: CGFloat
     let velocity: CGPoint
     let targetNotchIndex: Int
-    let targetNotchHeight: CGFloat
+    let targetTranslationHeight: CGFloat
     let notchHeightByIndex: [Int: CGFloat]
     let reachableIndexes: [Int]
 
@@ -25,5 +27,9 @@ struct ConcreteOverlayContainerContextTransitioning:
 
     func height(forNotchAt index: Int) -> CGFloat {
         return notchHeightByIndex[index] ?? 0
+    }
+
+    var targetNotchHeight: CGFloat {
+        return targetTranslationHeight
     }
 }