+++ Source/Classes/OverlayContainerTransitionCoordinator.swift
@@ -7,13 +7,16 @@
 
 import UIKit
 
-/// A protocol that provides information about an in-progress translation end.
+/// A protocol that provides information about an in-progress translation.
 /// Do not adopt this protocol in your own classes. Use the one provided by the `OverlayContainerTransitionCoordinator`.
 public protocol OverlayContainerTransitionCoordinatorContext {
-    /// The notch's index the container expects to reach.
-    var targetNotchIndex: Int { get }
+
+    /// A Boolean value indicating whether the transition is explicitly animated.
+    var isAnimated: Bool { get }
+    /// A Boolean value indicating whether the transition was cancelled.
+    var isCancelled: Bool { get }
     /// The notch's height the container expects to reach.
-    var targetNotchHeight: CGFloat { get }
+    var targetTranslationHeight: CGFloat { get }
     /// The current translation height.
     var overlayTranslationHeight: CGFloat { get }
     /// The notch indexes.
@@ -24,11 +27,10 @@ public protocol OverlayContainerTransitionCoordinatorContext {
     func height(forNotchAt index: Int) -> CGFloat
 }
 
-/// A protocol that provides support for animations associated with a overlay translation end.
+/// A protocol that provides support for animations associated with a overlay translation.
 ///
 /// Do not adopt this procotol in your own classes. Use the one provided by the `OverlayContainerDelegate` to
-/// add any extra animations alongside the translation end animations. It occurs when the user finish dragging
-/// the container's child view controllers.
+/// add any extra animations alongside the translation animations.
 public protocol OverlayContainerTransitionCoordinator: OverlayContainerTransitionCoordinatorContext {
     /// Runs the specified animations at the same time as overlay translation end animations.
     func animate(alongsideTransition animation: ((OverlayContainerTransitionCoordinatorContext) -> Void)?,