+++ Tests/OverlayNotchHeightTest.swift
@@ -71,23 +71,29 @@ class OverlayNotchHeightTest: QuickSpec {
         var overlayContainer: OverlayContainerViewController!
         var delegate: OverlayContainerDelegateImplementation!
         var overlay: UIViewController!
+
         beforeEach {
             overlayContainer = OverlayContainerViewController()
             delegate = OverlayContainerDelegateImplementation()
             overlay = UIViewController()
-            overlayContainer.loadViewIfNeeded()
             overlayContainer.view.frame = CGRect(origin: .zero, size: CGSize(width: 300, height: 300))
+        }
+
+        it("should move the overlay to the specified notch") {
             overlayContainer.delegate = delegate
             overlayContainer.viewControllers = [overlay]
-        }
-        it("should move the overlay to the specified notch") {
+            overlayContainer.loadViewIfNeeded()
             Notch.allCases.forEach { notch in
                 overlayContainer.moveOverlay(toNotchAt: notch.rawValue, animated: false)
                 overlayContainer.view.layoutIfNeeded()
                 expect(overlay.view.frame.height).to(equal(delegate.height(for: notch)))
             }
         }
+
         it("should invalidate the current notches") {
+            overlayContainer.delegate = delegate
+            overlayContainer.viewControllers = [overlay]
+            overlayContainer.loadViewIfNeeded()
             Notch.allCases.forEach { notch in
                 overlayContainer.moveOverlay(toNotchAt: notch.rawValue, animated: false)
                 overlayContainer.view.layoutIfNeeded()
@@ -101,25 +107,75 @@ class OverlayNotchHeightTest: QuickSpec {
                 expect(overlay.view.frame.height).to(equal(delegate.height(forNew: notch)))
             }
         }
+
         it("should call completion callback after animation is finished") {
+            overlayContainer.delegate = delegate
+            overlayContainer.viewControllers = [overlay]
+            overlayContainer.loadViewIfNeeded()
+            overlayContainer.view.layoutIfNeeded()
             Notch.allCases.forEach { notch in
-                waitUntil(timeout: 5) { done in
+                waitUntil(timeout: 1) { done in
                     overlayContainer.moveOverlay(toNotchAt: notch.rawValue, animated: true) {
                         expect(true).to(beTrue())
                         done()
                     }
+                    overlayContainer.view.layoutIfNeeded()
                 }
             }
         }
+
         it("should call completion callback when animation is set to false") {
+            overlayContainer.delegate = delegate
+            overlayContainer.viewControllers = [overlay]
+            overlayContainer.loadViewIfNeeded()
+            overlayContainer.view.layoutIfNeeded()
             Notch.allCases.forEach { notch in
-                waitUntil(timeout: 5) { done in
+                waitUntil(timeout: 1) { done in
                     overlayContainer.moveOverlay(toNotchAt: notch.rawValue, animated: false) {
                         expect(true).to(beTrue())
                         done()
                     }
+                    overlayContainer.view.layoutIfNeeded()
                 }
             }
+        }
+
+        it("should call completion callbacks in order") {
+            overlayContainer.delegate = delegate
+            overlayContainer.viewControllers = [overlay]
+            overlayContainer.loadViewIfNeeded()
+            overlayContainer.view.layoutIfNeeded()
+            waitUntil(timeout: 1) { done in
+                var i = 0
+                overlayContainer.moveOverlay(toNotchAt: Notch.minimum.rawValue, animated: false) {
+                    i += 1
+                }
+                overlayContainer.moveOverlay(toNotchAt: Notch.medium.rawValue, animated: false) {
+                    i += 1
+                    expect(i).to(equal(2))
+                    expect(overlay.view.frame.height).to(equal(delegate.height(for: .medium)))
+                    done()
+                }
+                overlayContainer.view.layoutIfNeeded()
+            }
+        }
+
+        it("should move the overlay even if the overlay is set before the delegate") {
+            let notch = Notch.maximum
+            overlayContainer.delegate = delegate
+            overlayContainer.moveOverlay(toNotchAt: notch.rawValue, animated: false)
+            overlayContainer.viewControllers = [overlay]
+            overlayContainer.view.layoutIfNeeded()
+            expect(overlay.view.frame.height).to(equal(delegate.height(for: notch)))
+        }
+
+        it("should move the overlay even if the overlay is set after the delegate") {
+            let notch = Notch.maximum
+            overlayContainer.viewControllers = [overlay]
+            overlayContainer.delegate = delegate
+            overlayContainer.moveOverlay(toNotchAt: notch.rawValue, animated: false)
+            overlayContainer.view.layoutIfNeeded()
+            expect(overlay.view.frame.height).to(equal(delegate.height(for: notch)))
         }
     }
 }